Enemy can damage player

+++ Assets/Script/EnnemyStat.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class EnnemyStat : MonoBehaviour
 {
     [SerializeField] public int maxHealth;
-    [SerializeField] public float damage;
+    [SerializeField] public int damage;
     [SerializeField] public float moveSpeed;
     [SerializeField] public float detectionRange;
     [SerializeField] public float chaseDistance;
